ANT-4633 - parquet : remove not needed static casts

+++ src/tests/src/libs/antares/writer/test-write-parquet-standalone.cpp
@@ -127,12 +127,12 @@ BOOST_AUTO_TEST_CASE(make_and_write_on_disk_a_regular_table_in_parquet___output_
     // chunk_size=1024 and we only have 3 rows: single row group, single chunk per column
     auto read_table = readParquet(file_path);
 
-    BOOST_REQUIRE_EQUAL(read_table->num_rows(), static_cast<int64_t>(ids.size()));
+    BOOST_REQUIRE_EQUAL(read_table->num_rows(), ids.size());
     BOOST_REQUIRE_EQUAL(read_table->num_columns(), 3);
 
     // Check "id" column (Int32)
     auto id_col = std::static_pointer_cast<arrow::Int32Array>(read_table->column(0)->chunk(0));
-    for (int i = 0; i < static_cast<int>(ids.size()); ++i)
+    for (int i = 0; i < ids.size(); ++i)
     {
         BOOST_CHECK(!id_col->IsNull(i));
         BOOST_CHECK_EQUAL(id_col->Value(i), ids[i]);
@@ -140,7 +140,7 @@ BOOST_AUTO_TEST_CASE(make_and_write_on_disk_a_regular_table_in_parquet___output_
 
     // Check "name" column (Utf8/String)
     auto name_col = std::static_pointer_cast<arrow::StringArray>(read_table->column(1)->chunk(0));
-    for (int i = 0; i < static_cast<int>(names.size()); ++i)
+    for (int i = 0; i < names.size(); ++i)
     {
         BOOST_CHECK(!name_col->IsNull(i));
         BOOST_CHECK_EQUAL(name_col->GetString(i), names[i]);
@@ -148,7 +148,7 @@ BOOST_AUTO_TEST_CASE(make_and_write_on_disk_a_regular_table_in_parquet___output_
 
     // Check "score" column (Double)
     auto score_col = std::static_pointer_cast<arrow::DoubleArray>(read_table->column(2)->chunk(0));
-    for (int i = 0; i < static_cast<int>(scores.size()); ++i)
+    for (int i = 0; i < scores.size(); ++i)
     {
         BOOST_CHECK(!score_col->IsNull(i));
         BOOST_CHECK_EQUAL(score_col->Value(i), scores[i]);
@@ -246,12 +246,12 @@ BOOST_AUTO_TEST_CASE(make_and_write_on_disk_a_nullable_table_in_parquet___output
     // chunk_size=1024 and we only have 3 rows: single row group, single chunk per column
     auto read_table = readParquet(file_path);
 
-    BOOST_REQUIRE_EQUAL(read_table->num_rows(), static_cast<int64_t>(ids.size()));
+    BOOST_REQUIRE_EQUAL(read_table->num_rows(), ids.size());
     BOOST_REQUIRE_EQUAL(read_table->num_columns(), 3);
 
     // Check "id" column — ids[0] is nullopt
     auto id_col = std::static_pointer_cast<arrow::Int32Array>(read_table->column(0)->chunk(0));
-    for (int i = 0; i < static_cast<int>(ids.size()); ++i)
+    for (int i = 0; i < ids.size(); ++i)
     {
         BOOST_CHECK_EQUAL(id_col->IsNull(i), !ids[i].has_value());
         if (ids[i].has_value())
@@ -262,7 +262,7 @@ BOOST_AUTO_TEST_CASE(make_and_write_on_disk_a_nullable_table_in_parquet___output
 
     // Check "name" column — names[1] is nullopt
     auto name_col = std::static_pointer_cast<arrow::StringArray>(read_table->column(1)->chunk(0));
-    for (int i = 0; i < static_cast<int>(names.size()); ++i)
+    for (int i = 0; i < names.size(); ++i)
     {
         BOOST_CHECK_EQUAL(name_col->IsNull(i), !names[i].has_value());
         if (names[i].has_value())
@@ -273,7 +273,7 @@ BOOST_AUTO_TEST_CASE(make_and_write_on_disk_a_nullable_table_in_parquet___output
 
     // Check "score" column — scores[2] is nullopt
     auto score_col = std::static_pointer_cast<arrow::DoubleArray>(read_table->column(2)->chunk(0));
-    for (int i = 0; i < static_cast<int>(scores.size()); ++i)
+    for (int i = 0; i < scores.size(); ++i)
     {
         BOOST_CHECK_EQUAL(score_col->IsNull(i), !scores[i].has_value());
         if (scores[i].has_value())
